refactor(gtk-host): the scroll edge is not React's
`onScrollNearEnd` stayed in `@gjsify/react-native`'s list controller when the rest
of the neutral half moved out, and review was right that it should not have: 34
lines of `Gtk.Adjustment` arithmetic, a measured claim about where GTK keeps a
scroll position, no React and no toolkit value import. Vue and Solid lists need
the same scroll-edge detection and would have re-measured it — which is the exact
failure this extraction exists to prevent, one file over from where it was being
prevented.

One policy in it IS React Native's: firing once per arrival at the end rather than
on every frame, which is `onEndReached`'s semantics. It moves with the geometry
because every "load more" caller wants it — firing per frame while the user rests
at the bottom would issue a page request per frame — and the header now says so
instead of leaving it unattributed. A dialect wanting the raw stream has the
adjustment and can subscribe to it directly.

`@gjsify/react-native` re-exports it, so `components.ts` and any consumer are
unchanged.

Also here, because it is the same file: `ListRowSink.mountRow`'s refusal note named
the LATER of the two routes to `occupied-slot`. GTK recycling a carrier across
bind/unbind is the one people expect and it needs scrolling to happen at all; the
route that fires first is `setRows` over unchanged keys, which re-shows a live row
with no GTK involvement — which is what the in-place vector already exercises.

react-native/src/lists/controller.ts: 227 -> 188 lines (389 before the extraction).
